refactor(client): step 5 -- peel storage families + finalize composition; monolith gone (#520)
Pure movement, final step. Extracts the two storage families into mixins
over _CoreClient:

  client/storage_tables.py  _StorageTablesMixin  (35 methods: buckets, tables,
                            snapshots, bucket sharing/linking)  1,164 LOC
  client/storage_files.py   _StorageFilesMixin   (10 methods: upload/download
                            incl. sliced, tagging)               342 LOC

_client.py is now a pure composition module: KeboolaClient assembled from all
ten mixins over _CoreClient, in the order named in issue #520, keeping its
class docstring. The residual monolith (all method bodies + their imports) is
gone -- every method now lives in exactly one family module and the stale
`logger`/imports left behind were dropped.

Judgment call: _StorageFilesMixin.upload_file calls self.prepare_file_upload
and self._upload_to_cloud (both _StorageTablesMixin methods -- the shared
cloud-upload path used by upload_table too). As with get_oauth_url in step 3,
its `self` is annotated `self: "KeboolaClient"` so ty resolves the cross-
family call; runtime is unchanged.

Every client/ module is now under the 2,000-LOC hard ceiling (largest:
storage_tables.py at 1,164). make check green: 4648 passed, 8 skipped. ty
clean (only the 3 pre-existing unrelated warnings).
